modulated files, improved prompting by clunking code
- Extension folder is now modulated for easy maintenance and feature implementation
- The code is now send to API in the form of a Array where every item in the Array is a line of code. This will enhance the prediction of start_char and end_char by Gemini
