+++ server.py
@@ -4,22 +4,21 @@ import json
 
 app = FastAPI()
 
-
 @app.post("/analyze")
 async def analyze(request: Request):
     data = await request.json()
-    code = data["code"]
+    lines = data["lines"]  # Get the lines from the request data
 
     analyzer = CodeAnalyzer()
-    analyzer._input_code = code
-    style_guide = analyzer.analyze_code()
-    processed_style_guide = json.loads(style_guide)
-    issues = {"issues": processed_style_guide}
+    style_guide = analyzer.analyze_code(lines)
+
     if style_guide:
+        processed_style_guide = json.loads(style_guide)
+        issues = {"issues": processed_style_guide}
         return {"style_guide": issues}
     else:
         return {"error": "Analysis failed"}, 400
 
-
 if __name__ == "__main__":
-    app.run(port=5000)
+    import uvicorn
+    uvicorn.run(app, host="0.0.0.0", port=5000)
