+++ vscode-extension/src/logger.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode';
+
+export function getLogger(module: string): vscode.OutputChannel {
+  return vscode.window.createOutputChannel(`[${module}]`);
+}
