+++ vscode-extension/src/utils.ts
@@ -1,0 +1,16 @@
+import * as vscode from 'vscode';
+
+export function getSeverity(severity: string): vscode.DiagnosticSeverity {
+  switch (severity.toUpperCase()) {
+    case 'ERROR':
+      return vscode.DiagnosticSeverity.Error;
+    case 'WARNING':
+      return vscode.DiagnosticSeverity.Warning;
+    case 'INFORMATION':
+      return vscode.DiagnosticSeverity.Information;
+    case 'HINT':
+      return vscode.DiagnosticSeverity.Hint;
+    default:
+      return vscode.DiagnosticSeverity.Information;
+  }
+}
